refactor: nwd module
