+++ src/modules/commands/nwd/ls/helpers/sortDirents.js
